refactor: direct_or_shell_command
